requirements_gpu.txt
tensorflow-gpu==1.12.0
keras==2.1.6
numpy==1.16.2
matplotlib
pillow